+++ examples/run_py.sh
@@ -1,6 +1,0 @@
-# python -u train.py 2>&1 | tee output.txt
-python -u train.py 5 ;\
-python -u train.py 10 ;\
-python -u train.py 20 ;\
-python -u train.py 40 ;\
-python -u train.py 50